refactored plot into its own component
