+++ package.json
@@ -28,7 +28,7 @@
     "markdown-it-synapse-table": "^1.0.6",
     "node-sass-chokidar": "^1.3.3",
     "npm-run-all": "^4.1.3",
-    "plotly.js": "^1.39.4",
+    "plotly.js": "^1.40.0",
     "prop-types": "^15.6.2",
     "react-plotly.js": "^2.2.0",
     "sanitize-html": "^1.18.4"
