+++ src/lib/containers/MarkdownSynapse.js
@@ -1,6 +1,8 @@
 import React from "react";
-import * as SynapseConstants from '../utils/SynapseConstants'
+import ReactDOM from 'react-dom'
+
 import * as SynapseClient from '../utils/SynapseClient'
+import SynapsePlot from './widgets/SynapsePlot'
 
 import PropTypes from 'prop-types'
 
@@ -84,7 +86,6 @@ class MarkdownSynapse extends React.Component {
 
         // handling each of the synapse widgets
         this.handleImageWidget = this.handleImageWidget.bind(this)
-        this.handlePlotlyWidget = this.handlePlotlyWidget.bind(this)
         this.handleReferenceWidget = this.handleReferenceWidget.bind(this)
         this.addBookmarks = this.addBookmarks.bind(this)
     }
@@ -141,7 +142,6 @@ class MarkdownSynapse extends React.Component {
     async processWidgetMappings(widgets, fileHandleAssociationList, elementList) {
         for (let element of widgets) {
             let widgetstring = element.dataset.widgetparams;
-            window.Elememeem = element
             let questionIndex = widgetstring.indexOf("?"); // type?
             let widgetType = widgetstring.substring(0, questionIndex); // type
             let widgetparamsMapped = {};
@@ -244,7 +244,12 @@ class MarkdownSynapse extends React.Component {
                 })
             }
         } else if (widgetType === "plot") {
-            elementList.push({element:element, widgetType: widgetType, widgetparamsMapped: widgetparamsMapped});
+            let plotWidget = <SynapsePlot 
+                                token={this.props.token}
+                                ownerId={this.props.ownerId}
+                                wikiId={this.props.wikiId}
+                                widgetparamsMapped={widgetparamsMapped} />
+            ReactDOM.render(plotWidget, element)
         } else if (widgetType === "reference") {
             elementList.push({element:element, widgetType: widgetType, widgetparamsMapped: widgetparamsMapped});
         }
@@ -266,8 +271,6 @@ class MarkdownSynapse extends React.Component {
                     // match corresponds to filehandle that this current element needs to be connected to
                     let match = this.matchToHandle(this.compareById(elementBundle.id, "fileHandleId"), this.state.fileResults);
                     this.handleImageWidget(match, elementBundle);
-                } else if (elementBundle.widgetType === "plot") {
-                    this.handlePlotlyWidget(elementBundle);
                 } else if (elementBundle.widgetType === "reference") {
                     this.handleReferenceWidget(elementBundle, referenceCount++, savedReferences);
                 }
@@ -283,109 +286,7 @@ class MarkdownSynapse extends React.Component {
             let renderedHTML = `<image class="img-fluid" src=${match[0].preSignedURL}></image>`;
             elementBundle.element.outerHTML = renderedHTML;
         }
-    }
-
-    handlePlotlyWidget(elementBundle) {
-        let widgetparamsMapped = elementBundle.widgetparamsMapped;
-        let raw_plot_data = null;
-        if (!this.state.queryData || !this.state.queryData[widgetparamsMapped.query]) {
-            // grab all the data, hasn't been loaded yet
-            raw_plot_data = this.getPlotlyData(widgetparamsMapped);
-        }
-        else {
-            // data already exists, don't regenerate
-            raw_plot_data = this.state.queryData[widgetparamsMapped.query];
-        }
-        if (!raw_plot_data) {
-            return
-        }
-        // grab all the parameters passed into the widget
-        let title = widgetparamsMapped.title;
-        let xtitle = widgetparamsMapped.xtitle;
-        let ytitle = widgetparamsMapped.ytitle;
-        let type = widgetparamsMapped.type;
-        let xaxisType = widgetparamsMapped.xaxistype;
-        let isHorizontal = widgetparamsMapped.horizontal.toLowerCase();
-        let showLegend = widgetparamsMapped.showlegend;
-        let layout = {
-            title: title,
-            showlegend: showLegend,
-        };
-        if (xtitle) {
-            layout.xaxis = {
-                title: xtitle
-            };
-        }
-        if (xaxisType) {
-            layout.xaxis = {
-                ...layout.xaxis,
-                xaxistype: xaxisType.toLowerCase()
-            };
-        }
-        if (ytitle) {
-            layout.yaxis = {
-                title: ytitle
-            };
-        }
-        let config = {
-            displayModeBar: false
-        };
-        if (!raw_plot_data.queryResult) {
-            // results haven't loaded yet
-            return null
-        }
-        // init plot_data
-        let plot_data = [];
-        let orientation = isHorizontal ? "v" : "h";
-        let headers = raw_plot_data.queryResult.queryResults.headers;
-        for (let i = 0; i < headers.length - 1; i++) {
-            // make an entry for each set of data points
-            plot_data[i] = {
-                x: [],
-                y: [],
-                name: headers[i + 1].name,
-                type: type.toLowerCase(),
-                orientation: orientation
-            };
-        }
-        // grab all the data
-        for (let i = 0; i < raw_plot_data.queryResult.queryResults.rows.length; i++) {
-            let row = raw_plot_data.queryResult.queryResults.rows[i];
-            for (let j = 1; j < row.values.length; j++) {
-                // create pairs of data
-                let row_values = row.values;
-                plot_data[j - 1].x.push(row_values[0]);
-                plot_data[j - 1].y.push(row_values[j]);
-            }
-        }
-        // error with clearing html - "" is not a function 
-        // wrapping in try/catch prevents the error, although it doesn't catch it.
-        try {
-            elementBundle.element.innerHTML = ""; // clear formatting (e.g. <Synapse Widget></SynapseWidget>)
-        }
-        catch (e) {
-            console.log('element bundle error ', e);
-        }
-        // responsive plot
-        // https://plot.ly/javascript/responsive-fluid-layout/#responsive--fluid-layout
-        (function () {
-            var d3 = window.Plotly.d3;
-            var WIDTH_IN_PERCENT_OF_PARENT = 100, HEIGHT_IN_PERCENT_OF_PARENT = 75;
-            var gd3 = d3.select(elementBundle.element)
-                .append('div')
-                .style({
-                    width: WIDTH_IN_PERCENT_OF_PARENT + '%',
-                    'margin-left': (100 - WIDTH_IN_PERCENT_OF_PARENT) / 5 + '%',
-                    height: HEIGHT_IN_PERCENT_OF_PARENT + 'vh',
-                    'margin-top': (100 - HEIGHT_IN_PERCENT_OF_PARENT) / 5 + 'vh'
-                });
-            var gd = gd3.node();
-            window.Plotly.plot(gd, plot_data, layout, config);
-            window.onresize = function () {
-                window.Plotly.Plots.resize(gd);
-            };
-        })();
-    }                
+    }         
 
     /**
      * Handles ?{reference} synapse widgets
@@ -491,87 +392,6 @@ class MarkdownSynapse extends React.Component {
     }
 
     /**
-     * Get data for plotly
-     *
-     * @param {*} widgetparamsMapped
-     * @returns data corresponding to plotly widget
-     * @memberof Markdown
-     */
-    getPlotlyData(widgetparamsMapped) {
-
-        let raw_plot_data = {}
-        let maxPageSize = 150
-
-        // step 1: get init query with maxRowsPerPage calculated
-        let queryRequest = {
-            concreteType: "org.sagebionetworks.repo.model.table.QueryBundleRequest",
-            entityId: this.props.ownerId,
-            query: {
-                isConsistent: false,
-                limit: maxPageSize,
-                partMask: SynapseConstants.BUNDLE_MASK_QUERY_RESULTS | SynapseConstants.BUNDLE_MASK_QUERY_FACETS, // 9,  // get query results and max rows per page
-                offset: 0,
-                sql: widgetparamsMapped.query
-            }
-        };
-
-        // Have to make two "sets" of calls for query, the first one tells us the maximum size per page of data
-        // we can get, the following uses that maximum and offsets to the appropriate location to get the data
-        // afterwards, the process repeats
-        SynapseClient.getQueryTableResults(queryRequest, this.props.token).then(initData => {
-            let queryCount = initData.queryResult.queryResults.rows.length
-            let totalQueryResults = queryCount
-            raw_plot_data = initData;
-            // Get the subsequent data, note- although the function calls itself, it runs
-            // iteratively due to the await
-            const getData = async () => {
-                if (queryCount === maxPageSize) {
-                    maxPageSize = initData.maxRowsPerPage
-                    let queryRequestWithMaxPageSize = {
-                        concreteType: "org.sagebionetworks.repo.model.table.QueryBundleRequest",
-                        entityId: this.props.ownerId,
-                        partMask: SynapseConstants.BUNDLE_MASK_QUERY_RESULTS,
-                        query: {
-                            isConsistent: false,
-                            limit: maxPageSize,
-                            offset: totalQueryResults,
-                            sql: widgetparamsMapped.query
-                        }
-                    };
-                    await SynapseClient.getQueryTableResults(queryRequestWithMaxPageSize, this.props.token)
-                        .then(post_data => {
-                            queryCount += post_data.queryResult.queryResults.rows.length
-                            if (queryCount > 0) {
-                                totalQueryResults += queryCount
-                                raw_plot_data.queryResult.queryResults.rows.push(
-                                    ...post_data.queryResult.queryResults.rows  // ... spread operator to push all elements on
-                                )
-                            }
-                            return getData()
-                        })
-                        .catch(err => 
-                            {
-                                console.log("Error on getting table results ", err)
-                            }
-                        );
-                } else {
-                    // set data to this plots sql in the query data
-                    let queryData = { ...this.state.queryData }; // shallow copy
-                    let query = widgetparamsMapped.query
-                    queryData[query] = raw_plot_data;
-                    this.setState({
-                        queryData
-                    });
-                    return raw_plot_data
-                }
-            }
-            return getData()
-        });
-        // when data hasn't loaded yet
-        return null
-    }
-
-    /**
      * Attach markdown to wiki attachments
      */
     matchToHandle(comparator, objectList) {
@@ -615,7 +435,7 @@ class MarkdownSynapse extends React.Component {
         // else the wiki page was retrieved accordingly or it was passed down
         // as a prop
     }
-
+              
     /**
      * Call Synapse REST API to get AMP-AD wiki portal attachments
      */
